Change examples package names
